2012-04-05 Roman - FIX:DAQ.Template: Restore template build after module DAQ.JavaLikeCalc or builded function close. - FIX:Special.SystemTests: Original message fixed. - FIX:DAQ.OPC_UA: Protocol requests counter show is fixed.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1781 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/OPC_UA/mod_prt.cpp
@@ -2692,7 +2692,7 @@ string OPCEndPoint::getStatus( )
     if( enableStat( ) )
     {
 	rez = _("Enabled. ");
-	rez += TSYS::strMess( _("Requests %.4g."), cntReq );
+	rez += TSYS::strMess( _("Requests %.4g."), (double)cntReq );
     }
 
     return rez;
